+++ src/containers/my-resources/add-attachment-category-modal/AddAttachmentCategoryModal.tsx
@@ -4,7 +4,7 @@ import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 
 import AppButton from '~/components/app-button/AppButton'
-import AsyncAutocomplete from '~/components/async-autocomlete/AsyncAutocomplete'
+import AsyncAutocomplete from '~/components/async-autocomplete/AsyncAutocomplete'
 
 import { ResourceService } from '~/services/resource-service'
 import useForm from '~/hooks/use-form'
